restructured 'mock-data/mock.ts' file: creted 'mock-data/offers.ts'
